+++ src/components/CartItem.tsx
@@ -1,7 +1,8 @@
 import { FC } from 'react';
 
-import { addItem, CartItemType, minusItem, removeItem } from '../redux/slices/cartSlice';
 import { useAppDispatch } from '../redux/store';
+import { addItem, minusItem, removeItem } from '../redux/cart/slice';
+import { CartItemType } from '../redux/cart/types';
 
 type CartItemProps = {
   id: string;
