+++ src/redux/filter/slice.ts
@@ -1,26 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../store';
-
-export enum SortPropertyEnum {
-  RATING_DESC = 'rating',
-  RATING_ASC = '-rating',
-  TITLE_DESC = 'title',
-  TITLE_ASC = '-title',
-  PRICE_DESC = 'price',
-  PRICE_ASC = '-price',
-}
-
-export type SortType = {
-  name: string;
-  sortProperty: SortPropertyEnum;
-};
-
-export interface FilterSliceState {
-  searchValue: string;
-  categoryId: number;
-  currentPage: number;
-  sort: SortType;
-}
+import { FilterSliceState, SortPropertyEnum, SortType } from './types';
 
 const initialState: FilterSliceState = {
   searchValue: '',
@@ -64,9 +43,6 @@ export const filterSlice = createSlice({
     },
   },
 });
-
-export const selectSort = (state: RootState) => state.filters.sort;
-export const selectFilter = (state: RootState) => state.filters;
 
 export const { setCategoryId, setSearchValue, setSort, setCurrentPage, setFilters } =
   filterSlice.actions;
