+++ src/redux/store.ts
@@ -1,9 +1,9 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { useDispatch } from 'react-redux';
 
-import filters from './slices/filterSlice';
-import cart from './slices/cartSlice';
-import pizza from './slices/pizzaSlice';
+import filters from './filter/slice';
+import cart from './cart/slice';
+import pizza from './pizza/slice';
 
 export const store = configureStore({
   reducer: {
