+++ src/utils/calcTotalPrice.ts
@@ -1,4 +1,4 @@
-import { CartItemType } from '../redux/slices/cartSlice';
+import { CartItemType } from '../redux/cart/types';
 
 export const calcTotalPrice = (items: CartItemType[]) => {
   return items.reduce((sum, obj) => obj.price * obj.count + sum, 0);
